Added "Share Link" Under Course Name
Created a "Share Course" text underneath each course name that copies the direct course link to the clipboard. Could obviously be better formatted, but I'm not too great at that part. As always, I barely touch JS and this was cobbled together from existing code and StackOverflow so if anything's horribly wrong feel free to fix it